split persisted expansion state into new file
